cloudrun/actas/requirements.txt
PyMuPDF==1.24.10
Pillow==10.4.0
pyzbar==0.1.9
google-cloud-storage==2.18.2
google-cloud-firestore==2.17.2
gunicorn==22.0.0
